Add the ability to set WP_HOME
This allows for a user to override the value in the wp_options table
(useful for development instances) as well as reduce database calls on
production instances. This also applies to wp_site_url and so I have
added this use to the description for that parameter.

I have added in some whitespace cleanup for the wp_site_url and wp_home
tags in the wp-config.erb.php - it may be desirable to replicate this
with all of the if/end statements in both template files to avoid
unnecessary whitespace.